moved server response handling from client into separate classes, small changes
