+++ src/main/java/ru/kpfu/itis/paramonov/heartstone/controller/BattlefieldController.java
@@ -4,7 +4,6 @@ import javafx.collections.ObservableList;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.Node;
-import javafx.scene.control.Button;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.*;
